✨ [#35] Separate ColorSelector component
ColorSelector 컴포넌트 Palette에서 분리
세로모드시 Tool 컴포넌트에 추가
